Fixes rebalancing bugs and adds random tests
The first bug caused rebalancing not to be performed when the
tallest child of the unbalanced node was balanced. This affected
both insertions and deletions.

The second bug caused the rebalancing of the child node to be
ignored by the parent, when the rebalancing was the result of
deletion of a node with two children.

Two random tests were added to test the height of the tree after
a number of insertions and to test the integrity of the tree after
a number of insertions and deletions.